requirements.txt
pandas==2.2.3
Pillow==11.1.0
qrcode==8.0
streamlit==1.42.2
streamlit==1.39.0
